refactor: split ComputeView into sub-components
Extract NodeTable, NodeAlerts, StatCard/UsageBar, MachineManagement,
MachineConfigSection, and shared types from ComputeView.tsx (1213 -> 697
lines). Health audit section left intact for separate extraction.
