+++ src/kubeview/views/ComputeView.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
 import { useQuery } from '@tanstack/react-query';
-import {
-  Server, Cpu, HardDrive, CheckCircle, XCircle, AlertCircle,
-  ArrowRight, Ban, Box, Terminal, FileText, Activity, Info, AlertTriangle,
-} from 'lucide-react';
+import { Server, CheckCircle, Activity, AlertTriangle } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { k8sList } from '../engine/query';
 import { queryInstant } from '../components/metrics/prometheus';
@@ -11,7 +8,7 @@ import { MetricCard } from '../components/metrics/Sparkline';
 import { CHART_COLORS } from '../engine/colors';
 import { MetricGrid } from '../components/primitives/MetricGrid';
 import type { K8sResource } from '../engine/renderers';
-import type { Node, Pod, Machine, MachineSet, Taint, NodePool, Condition } from '../engine/types';
+import type { Node, Pod, Machine, NodePool, Condition } from '../engine/types';
 import { getNodeStatus } from '../engine/renderers/statusUtils';
 import { parseResourceValue, formatBytes, formatCpu } from '../engine/formatting';
 import { useNavigateTab } from '../hooks/useNavigateTab';
@@ -19,54 +16,12 @@ import { useK8sListWatch } from '../hooks/useK8sListWatch';
 import { useClusterStore } from '../store/clusterStore';
 import { Card } from '../components/primitives/Card';
 import { CapacityTab } from './compute/CapacityTab';
-
-/** Prometheus instant-query result entry */
-type PrometheusResult = { metric: Record<string, string>; value: number };
-
-/** MachineAutoscaler shape (not in typed interfaces yet) */
-interface MachineAutoscalerResource extends K8sResource {
-  spec?: { scaleTargetRef?: { apiVersion?: string; kind?: string; name?: string }; minReplicas?: number; maxReplicas?: number };
-}
-
-/** ClusterAutoscaler shape (not in typed interfaces yet) */
-interface ClusterAutoscalerResource extends K8sResource {
-  spec?: { resourceLimits?: { maxNodesTotal?: number; minNodesTotal?: number }; scaleDown?: { enabled?: boolean; delayAfterAdd?: string } };
-}
-
-/** MachineHealthCheck shape (not in typed interfaces yet) */
-interface MachineHealthCheckResource extends K8sResource {
-  spec?: { maxUnhealthy?: string; unhealthyConditions?: Array<{ type: string; status: string; timeout: string }>; selector?: Record<string, unknown> };
-  status?: { currentHealthy?: number; expectedMachines?: number };
-}
-
-/** MachineConfigPool shape (not in typed interfaces yet) */
-interface MachineConfigPoolResource extends K8sResource {
-  status?: { conditions?: Array<{ type: string; status: string; reason?: string; message?: string }>; machineCount?: number; readyMachineCount?: number; updatedMachineCount?: number; configuration?: { name?: string } };
-}
-
-/** Per-node detail computed for the table & audit checks */
-interface NodeDetail {
-  node: K8sResource;
-  status: ReturnType<typeof getNodeStatus>;
-  nodeInfo: Partial<NonNullable<NonNullable<Node['status']>['nodeInfo']>>;
-  capacity: Partial<NonNullable<NonNullable<Node['status']>['capacity']>>;
-  allocatable: Partial<NonNullable<NonNullable<Node['status']>['capacity']>>;
-  roles: string[];
-  taints: NonNullable<NonNullable<Node['spec']>['taints']>;
-  unschedulable: boolean | undefined;
-  podCount: number;
-  podCap: number;
-  cpuCap: number;
-  memCap: number;
-  memUsagePct: number | null;
-  cpuUsagePct: number | null;
-  age: string;
-  pressures: string[];
-  instanceType: string;
-  name: string;
-}
-
-// parseQuantity, formatBytes, formatCpu imported from engine/formatting.ts
+import { StatCard } from './compute/StatCard';
+import { NodeTable } from './compute/NodeTable';
+import { NodeAlerts } from './compute/NodeAlerts';
+import { MachineManagement } from './compute/MachineManagement';
+import { MachineConfigSection } from './compute/MachineConfigSection';
+import type { NodeDetail, PrometheusResult } from './compute/types';
 
 export default function ComputeView() {
   const go = useNavigateTab();
@@ -318,498 +273,27 @@ export default function ComputeView() {
         />
 
         {/* Alerts */}
-        {(unreadyNodes.length > 0 || pressureNodes.length > 0) && (
-          <div className="space-y-2">
-            {unreadyNodes.map((n) => (
-              <button key={n.metadata.uid} onClick={() => go(`/r/v1~nodes/_/${n.metadata.name}`, n.metadata.name)}
-                className="w-full flex items-center gap-3 px-4 py-3 bg-red-950/30 border border-red-900 rounded-lg hover:bg-red-950/50 text-left">
-                <XCircle className="w-5 h-5 text-red-400 shrink-0" />
-                <div className="flex-1"><span className="text-sm font-medium text-slate-200">{n.metadata.name}</span><span className="text-xs text-red-400 ml-2">NotReady</span></div>
-                <ArrowRight className="w-4 h-4 text-slate-600" />
-              </button>
-            ))}
-            {pressureNodes.map((n) => {
-              const s = getNodeStatus(n);
-              const pressures = [s.pressure.disk && 'Disk', s.pressure.memory && 'Memory', s.pressure.pid && 'PID'].filter(Boolean).join(', ');
-              return (
-                <button key={n.metadata.uid} onClick={() => go(`/r/v1~nodes/_/${n.metadata.name}`, n.metadata.name)}
-                  className="w-full flex items-center gap-3 px-4 py-3 bg-yellow-950/30 border border-yellow-900 rounded-lg hover:bg-yellow-950/50 text-left">
-                  <AlertCircle className="w-5 h-5 text-yellow-400 shrink-0" />
-                  <div className="flex-1"><span className="text-sm font-medium text-slate-200">{n.metadata.name}</span><span className="text-xs text-yellow-400 ml-2">{pressures} Pressure</span></div>
-                  <ArrowRight className="w-4 h-4 text-slate-600" />
-                </button>
-              );
-            })}
-          </div>
-        )}
+        <NodeAlerts unreadyNodes={unreadyNodes} pressureNodes={pressureNodes} go={go} />
 
         {/* Node table */}
-        <Card className="overflow-hidden">
-          <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-            <h2 className="text-sm font-semibold text-slate-100">Nodes ({nodes.length})</h2>
-            <button onClick={() => go('/r/v1~nodes', 'Nodes')} className="text-xs text-blue-400 hover:text-blue-300">View all →</button>
-          </div>
+        <NodeTable nodeDetails={nodeDetails} totalCount={nodes.length} go={go} />
 
-          <div className="overflow-x-auto">
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="border-b border-slate-800 bg-slate-900/50">
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Node</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Status</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Roles</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">CPU</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Memory</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Pods</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Instance</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Version</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400">Age</th>
-                  <th className="px-4 py-2 text-left text-xs font-semibold text-slate-400"></th>
-                </tr>
-              </thead>
-              <tbody className="divide-y divide-slate-800">
-                {nodeDetails.map((nd) => {
-                  const podPct = nd.podCap > 0 ? (nd.podCount / nd.podCap) * 100 : 0;
-                  const memPct = nd.memUsagePct;
-                  const cpuPct = nd.cpuUsagePct;
-                  return (
-                    <tr key={nd.node.metadata.uid} onClick={() => go(`/r/v1~nodes/_/${nd.name}`, nd.name)}
-                      className="hover:bg-slate-800/70 cursor-pointer transition-colors">
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-2">
-                          {nd.status.ready ? <CheckCircle className="w-3.5 h-3.5 text-green-500 shrink-0" /> : <XCircle className="w-3.5 h-3.5 text-red-500 shrink-0" />}
-                          <span className="text-slate-200 truncate max-w-[200px]" title={nd.name}>{nd.name}</span>
-                          {nd.unschedulable && <Ban className="w-3 h-3 text-yellow-500" title="Cordoned" />}
-                        </div>
-                        {nd.taints.length > 0 && (
-                          <div className="flex flex-wrap gap-1 mt-1">
-                            {nd.taints.map((t, i) => (
-                              <span key={i} className={cn('text-xs px-1 py-0.5 rounded font-mono',
-                                t.effect === 'NoSchedule' ? 'bg-yellow-900/30 text-yellow-400' :
-                                t.effect === 'NoExecute' ? 'bg-red-900/30 text-red-400' :
-                                'bg-slate-800 text-slate-500'
-                              )} title={`${t.key}=${t.value || ''}:${t.effect}`}>
-                                {t.key.split('/').pop()}:{t.effect}
-                              </span>
-                            ))}
-                          </div>
-                        )}
-                      </td>
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-1.5">
-                          <span className={cn('text-xs px-1.5 py-0.5 rounded', nd.status.ready ? 'bg-green-900/50 text-green-300' : 'bg-red-900/50 text-red-300')}>
-                            {nd.status.ready ? 'Ready' : 'NotReady'}
-                          </span>
-                          {nd.pressures.length > 0 && nd.pressures.map((p: string) => (
-                            <span key={p} className="text-xs px-1 py-0.5 bg-red-900/50 text-red-300 rounded">{p}</span>
-                          ))}
-                        </div>
-                      </td>
-                      <td className="px-4 py-3">
-                        <div className="flex gap-1">{nd.roles.map((r: string) => <span key={r} className={cn('text-xs px-1.5 py-0.5 rounded', r === 'master' || r === 'control-plane' ? 'bg-purple-900/50 text-purple-300' : r === 'infra' ? 'bg-orange-900/50 text-orange-300' : 'bg-blue-900/50 text-blue-300')}>{r}</span>)}</div>
-                      </td>
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-2">
-                          {cpuPct !== null ? (
-                            <>
-                              <UsageBar pct={cpuPct} color={cpuPct > 80 ? 'red' : cpuPct > 60 ? 'yellow' : 'green'} />
-                              <span className="text-xs text-slate-400 font-mono w-10">{Math.round(cpuPct)}%</span>
-                            </>
-                          ) : (
-                            <span className="text-xs text-slate-500">{formatCpu(nd.cpuCap)}</span>
-                          )}
-                        </div>
-                      </td>
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-2">
-                          {memPct !== null ? (
-                            <>
-                              <UsageBar pct={memPct} color={memPct > 80 ? 'red' : memPct > 60 ? 'yellow' : 'green'} />
-                              <span className="text-xs text-slate-400 font-mono w-10">{Math.round(memPct)}%</span>
-                            </>
-                          ) : (
-                            <span className="text-xs text-slate-500">{formatBytes(nd.memCap)}</span>
-                          )}
-                        </div>
-                      </td>
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-2">
-                          <UsageBar pct={podPct} color={podPct > 80 ? 'red' : podPct > 60 ? 'yellow' : 'blue'} />
-                          <span className="text-xs text-slate-400 font-mono w-12">{nd.podCount}/{nd.podCap}</span>
-                        </div>
-                      </td>
-                      <td className="px-4 py-3">
-                        {nd.instanceType ? (
-                          <span className="text-xs px-1.5 py-0.5 bg-slate-800 text-slate-400 rounded font-mono">{nd.instanceType}</span>
-                        ) : (
-                          <span className="text-xs text-slate-600">{nd.nodeInfo.architecture}</span>
-                        )}
-                      </td>
-                      <td className="px-4 py-3"><span className="text-xs text-slate-500 font-mono">{nd.nodeInfo.kubeletVersion}</span></td>
-                      <td className="px-4 py-3"><span className="text-xs text-slate-500">{nd.age}</span></td>
-                      <td className="px-4 py-3">
-                        <div className="flex items-center gap-1">
-                          <button onClick={(e) => { e.stopPropagation(); go(`/node-logs/${nd.name}`, `${nd.name} (Logs)`); }} className="p-1 text-slate-500 hover:text-blue-400 transition-colors" title="Node Logs"><FileText className="w-3.5 h-3.5" /></button>
-                        </div>
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </div>
-        </Card>
-
-        {/* Machine Management — hidden on HyperShift where Machine API is managed externally */}
-        {isHyperShift ? (
-          <div className="space-y-6">
-            {/* HyperShift info banner */}
-            <div className="flex items-center gap-2 px-4 py-2.5 bg-blue-950/30 border border-blue-800/50 rounded-lg">
-              <Info className="w-4 h-4 text-blue-400 shrink-0" />
-              <span className="text-xs text-blue-300">Hosted Control Plane — Machine API managed externally. Worker scaling via NodePools.</span>
-            </div>
-
-            {/* NodePools */}
-            <Card>
-              <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-                <h2 className="text-sm font-semibold text-slate-100 flex items-center gap-2">
-                  NodePools ({nodePools.length})
-                  <span className="text-xs px-2 py-0.5 bg-violet-900/50 text-violet-300 rounded-full">HyperShift</span>
-                </h2>
-              </div>
-              {nodePools.length === 0 ? (
-                <div className="px-4 py-8 text-center">
-                  <p className="text-sm text-slate-500">No NodePools found</p>
-                  <p className="text-xs text-slate-600 mt-1">NodePool resources may not be accessible from the guest cluster, or the hypershift.openshift.io API is not available.</p>
-                </div>
-              ) : (
-                <div className="divide-y divide-slate-800">
-                  {(nodePools as unknown as NodePool[]).map((np) => {
-                    const desired = np.spec?.replicas ?? 0;
-                    const ready = np.status?.replicas ?? 0;
-                    const autoScale = np.spec?.autoScaling;
-                    const platform = np.spec?.platform;
-                    const instanceType = platform?.aws?.instanceType || platform?.azure?.vmSize || platform?.gcp?.instanceType || '';
-                    const autoRepair = np.spec?.management?.autoRepair;
-                    const version = np.status?.version;
-                    const conditions = np.status?.conditions || [];
-                    const isDegraded = conditions.some((c: Condition) => c.type === 'Ready' && c.status !== 'True');
-                    const isUpdating = conditions.some((c: Condition) => (c.type === 'UpdatingVersion' || c.type === 'UpdatingConfig') && c.status === 'True');
-
-                    return (
-                      <div key={np.metadata.uid || np.metadata.name} className="px-4 py-3 hover:bg-slate-800/30 transition-colors">
-                        <div className="flex items-center justify-between">
-                          <div className="flex items-center gap-2 min-w-0">
-                            <div className={cn('w-2 h-2 rounded-full shrink-0',
-                              isDegraded ? 'bg-red-500' : isUpdating ? 'bg-blue-500 animate-pulse' : ready === desired ? 'bg-green-500' : 'bg-yellow-500'
-                            )} />
-                            <span className="text-sm font-medium text-slate-200">{np.metadata.name}</span>
-                            {instanceType && <span className="text-xs px-1.5 py-0.5 bg-slate-800 text-slate-400 rounded font-mono">{instanceType}</span>}
-                            {autoScale && (
-                              <span className="text-xs px-1.5 py-0.5 bg-blue-900/50 text-blue-300 rounded">
-                                auto {autoScale.min}-{autoScale.max}
-                              </span>
-                            )}
-                            {autoRepair && <span className="text-xs px-1.5 py-0.5 bg-emerald-900/50 text-emerald-300 rounded">auto-repair</span>}
-                          </div>
-                          <div className="flex items-center gap-3 shrink-0">
-                            {version && <span className="text-xs text-slate-500 font-mono">{version}</span>}
-                            <span className={cn('text-xs font-mono font-semibold',
-                              ready === desired ? 'text-green-400' : 'text-yellow-400'
-                            )}>{ready}/{desired}</span>
-                          </div>
-                        </div>
-                        {isDegraded && (
-                          <div className="mt-1.5 ml-4">
-                            <span className="text-xs text-red-400">
-                              {conditions.find((c: Condition) => c.type === 'Ready' && c.status !== 'True')?.message || 'Not ready'}
-                            </span>
-                          </div>
-                        )}
-                      </div>
-                    );
-                  })}
-                </div>
-              )}
-            </Card>
-          </div>
-        ) : (
-        <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-          {/* MachineSets */}
-          <Card>
-            <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-              <h2 className="text-sm font-semibold text-slate-100">MachineSets ({machineSets.length})</h2>
-              <button onClick={() => go('/r/machine.openshift.io~v1beta1~machinesets', 'MachineSets')} className="text-xs text-blue-400 hover:text-blue-300">View all →</button>
-            </div>
-            <div className="divide-y divide-slate-800 max-h-64 overflow-auto">
-              {machineSets.length === 0 ? <div className="px-4 py-6 text-center text-sm text-slate-500">No MachineSets</div> : (machineSets as MachineSet[]).map((ms) => {
-                const desired = ms.spec?.replicas ?? 0;
-                const ready = ms.status?.readyReplicas ?? 0;
-                // Find autoscaler for this machineset
-                const autoscaler = (machineAutoscalers as MachineAutoscalerResource[]).find((a) => a.spec?.scaleTargetRef?.name === ms.metadata.name);
-                return (
-                  <button key={ms.metadata.uid} onClick={() => go(`/r/machine.openshift.io~v1beta1~machinesets/${ms.metadata.namespace}/${ms.metadata.name}`, ms.metadata.name)}
-                    className="w-full flex items-center justify-between px-4 py-2.5 hover:bg-slate-800/50 text-left">
-                    <div className="flex items-center gap-2 min-w-0">
-                      <div className={cn('w-2 h-2 rounded-full shrink-0', ready === desired ? 'bg-green-500' : 'bg-yellow-500')} />
-                      <span className="text-sm text-slate-200 truncate">{ms.metadata.name}</span>
-                      {autoscaler && <span className="text-xs px-1.5 py-0.5 bg-blue-900/50 text-blue-300 rounded shrink-0">autoscaled {autoscaler.spec?.minReplicas}-{autoscaler.spec?.maxReplicas}</span>}
-                    </div>
-                    <span className={cn('text-xs font-mono shrink-0', ready === desired ? 'text-green-400' : 'text-yellow-400')}>{ready}/{desired}</span>
-                  </button>
-                );
-              })}
-            </div>
-          </Card>
-
-          {/* Machines */}
-          <Card>
-            <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-              <h2 className="text-sm font-semibold text-slate-100">Machines ({machines.length})</h2>
-              <button onClick={() => go('/r/machine.openshift.io~v1beta1~machines', 'Machines')} className="text-xs text-blue-400 hover:text-blue-300">View all →</button>
-            </div>
-            <div className="divide-y divide-slate-800 max-h-64 overflow-auto">
-              {machines.length === 0 ? <div className="px-4 py-6 text-center text-sm text-slate-500">No Machines</div> : (machines as Machine[]).map((m) => {
-                const phase = m.status?.phase || 'Unknown';
-                const instanceType = (m.spec?.providerSpec?.value?.instanceType as string) || '';
-                const nodeRef = m.status?.nodeRef?.name || '';
-                return (
-                  <button key={m.metadata.uid} onClick={() => go(`/r/machine.openshift.io~v1beta1~machines/${m.metadata.namespace}/${m.metadata.name}`, m.metadata.name)}
-                    className="w-full flex items-center justify-between px-4 py-2.5 hover:bg-slate-800/50 text-left">
-                    <div className="flex items-center gap-2 min-w-0">
-                      <div className={cn('w-2 h-2 rounded-full shrink-0', phase === 'Running' ? 'bg-green-500' : phase === 'Provisioning' ? 'bg-yellow-500' : 'bg-red-500')} />
-                      <div className="min-w-0">
-                        <span className="text-sm text-slate-200 truncate block">{m.metadata.name}</span>
-                        {nodeRef && <span className="text-xs text-slate-500 truncate block">{nodeRef}</span>}
-                      </div>
-                    </div>
-                    <div className="flex items-center gap-2 shrink-0">
-                      {instanceType && <span className="text-xs px-1.5 py-0.5 bg-slate-800 text-slate-400 rounded font-mono">{instanceType}</span>}
-                      <span className={cn('text-xs px-1.5 py-0.5 rounded',
-                        phase === 'Running' ? 'bg-green-900/50 text-green-300' :
-                        phase === 'Provisioning' ? 'bg-yellow-900/50 text-yellow-300' :
-                        'bg-red-900/50 text-red-300'
-                      )}>{phase}</span>
-                    </div>
-                  </button>
-                );
-              })}
-            </div>
-          </Card>
-
-          {/* MachineHealthChecks */}
-          <Card>
-            <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-              <h2 className="text-sm font-semibold text-slate-100">Machine Health Checks ({healthChecks.length})</h2>
-              <button onClick={() => go('/r/machine.openshift.io~v1beta1~machinehealthchecks', 'HealthChecks')} className="text-xs text-blue-400 hover:text-blue-300">View all →</button>
-            </div>
-            <div className="divide-y divide-slate-800 max-h-64 overflow-auto">
-              {healthChecks.length === 0 ? <div className="px-4 py-6 text-center text-sm text-slate-500">No health checks configured</div> : (healthChecks as MachineHealthCheckResource[]).map((hc) => {
-                const maxUnhealthy = hc.spec?.maxUnhealthy || '100%';
-                const conditions = hc.spec?.unhealthyConditions || [];
-                const currentHealthy = hc.status?.currentHealthy ?? 0;
-                const expectedMachines = hc.status?.expectedMachines ?? 0;
-                return (
-                  <button key={hc.metadata.uid} onClick={() => go(`/r/machine.openshift.io~v1beta1~machinehealthchecks/${hc.metadata.namespace}/${hc.metadata.name}`, hc.metadata.name)}
-                    className="w-full px-4 py-3 hover:bg-slate-800/50 text-left">
-                    <div className="flex items-center justify-between mb-1">
-                      <span className="text-sm text-slate-200">{hc.metadata.name}</span>
-                      <span className={cn('text-xs font-mono', currentHealthy === expectedMachines ? 'text-green-400' : 'text-yellow-400')}>
-                        {currentHealthy}/{expectedMachines} healthy
-                      </span>
-                    </div>
-                    <div className="flex items-center gap-3 text-xs text-slate-500">
-                      <span>Max unhealthy: {maxUnhealthy}</span>
-                      {conditions.map((c, i) => <span key={i}>{c.type}≠{c.status} → {c.timeout}</span>)}
-                    </div>
-                  </button>
-                );
-              })}
-            </div>
-          </Card>
-
-          {/* Autoscaling */}
-          <Card>
-            <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-              <h2 className="text-sm font-semibold text-slate-100">Autoscaling</h2>
-              {clusterAutoscaler.length > 0 && <span className="text-xs px-1.5 py-0.5 bg-green-900/50 text-green-300 rounded">Enabled</span>}
-            </div>
-            <div className="p-4 space-y-3">
-              {clusterAutoscaler.length === 0 && machineAutoscalers.length === 0 ? (
-                <div className="py-3 space-y-4">
-                  <div className="text-sm text-slate-400">Autoscaling is not configured</div>
-
-                  <div className="space-y-3 text-xs text-slate-500">
-                    <div className="flex gap-3">
-                      <span className="text-blue-400 font-bold shrink-0">Step 1</span>
-                      <div>
-                        <div className="text-slate-300 font-medium">Create a ClusterAutoscaler</div>
-                        <div>Sets cluster-wide limits: max total nodes, max cores, max memory. Only one per cluster.</div>
-                      </div>
-                    </div>
-                    <div className="flex gap-3">
-                      <span className="text-blue-400 font-bold shrink-0">Step 2</span>
-                      <div>
-                        <div className="text-slate-300 font-medium">Create MachineAutoscalers (one per MachineSet)</div>
-                        <div>Sets min/max replicas per MachineSet. The autoscaler adds nodes when pods are pending due to insufficient resources, and removes them when utilization is low.</div>
-                      </div>
-                    </div>
-                  </div>
-
-                  <div className="flex items-center gap-3">
-                    <button onClick={() => go('/create/autoscaling.openshift.io~v1~clusterautoscalers', 'Create ClusterAutoscaler')} className="px-3 py-1.5 text-xs bg-blue-600 text-white rounded hover:bg-blue-500 flex items-center gap-1">
-                      Create ClusterAutoscaler <ArrowRight className="w-3 h-3" />
-                    </button>
-                    <button onClick={() => go('/create/autoscaling.openshift.io~v1beta1~machineautoscalers', 'Create MachineAutoscaler')} className="px-3 py-1.5 text-xs bg-slate-800 text-slate-200 rounded hover:bg-slate-700 flex items-center gap-1">
-                      Create MachineAutoscaler <ArrowRight className="w-3 h-3" />
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <>
-                  {(clusterAutoscaler as ClusterAutoscalerResource[]).map((ca) => (
-                    <div key={ca.metadata.uid} className="p-3 bg-slate-800/50 rounded border border-slate-700">
-                      <div className="text-sm text-slate-200 mb-2">Cluster Autoscaler</div>
-                      <div className="grid grid-cols-2 gap-2 text-xs">
-                        <div><span className="text-slate-500">Min nodes:</span> <span className="text-slate-300">{ca.spec?.resourceLimits?.minNodesTotal ?? '—'}</span></div>
-                        <div><span className="text-slate-500">Max nodes:</span> <span className="text-slate-300">{ca.spec?.resourceLimits?.maxNodesTotal ?? '—'}</span></div>
-                        <div><span className="text-slate-500">Scale down:</span> <span className="text-slate-300">{ca.spec?.scaleDown?.enabled ? 'enabled' : 'disabled'}</span></div>
-                        <div><span className="text-slate-500">Delay:</span> <span className="text-slate-300">{ca.spec?.scaleDown?.delayAfterAdd ?? '—'}</span></div>
-                      </div>
-                    </div>
-                  ))}
-                  {machineAutoscalers.length > 0 && (
-                    <div>
-                      <div className="text-xs text-slate-500 mb-2">Machine Autoscalers ({machineAutoscalers.length})</div>
-                      {(machineAutoscalers as MachineAutoscalerResource[]).map((ma) => (
-                        <div key={ma.metadata.uid} className="flex items-center justify-between py-1.5">
-                          <span className="text-sm text-slate-300">{ma.spec?.scaleTargetRef?.name || ma.metadata.name}</span>
-                          <span className="text-xs text-slate-400 font-mono">{ma.spec?.minReplicas}–{ma.spec?.maxReplicas} replicas</span>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                </>
-              )}
-            </div>
-          </Card>
-        </div>
-        )}
+        {/* Machine Management */}
+        <MachineManagement
+          isHyperShift={isHyperShift}
+          machines={machines}
+          machineSets={machineSets}
+          healthChecks={healthChecks}
+          machineAutoscalers={machineAutoscalers}
+          clusterAutoscaler={clusterAutoscaler}
+          nodePools={nodePools}
+          go={go}
+        />
 
         {/* MachineConfig Management */}
-        <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-          {/* MachineConfigPools */}
-          <Card>
-            <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-              <h2 className="text-sm font-semibold text-slate-100">MachineConfigPools ({machineConfigPools.length})</h2>
-              <button onClick={() => go('/r/machineconfiguration.openshift.io~v1~machineconfigpools', 'MachineConfigPools')} className="text-xs text-blue-400 hover:text-blue-300">View all →</button>
-            </div>
-            <div className="divide-y divide-slate-800 max-h-64 overflow-auto">
-              {machineConfigPools.length === 0 ? (
-                <div className="px-4 py-6 text-center text-sm text-slate-500">No MachineConfigPools found</div>
-              ) : (machineConfigPools as MachineConfigPoolResource[]).map((mcp) => {
-                const conditions = mcp.status?.conditions || [];
-                const updated = conditions.find((c) => c.type === 'Updated');
-                const updating = conditions.find((c) => c.type === 'Updating');
-                const degraded = conditions.find((c) => c.type === 'Degraded');
-                const isUpdated = updated?.status === 'True';
-                const isUpdating = updating?.status === 'True';
-                const isDegraded = degraded?.status === 'True';
-                const machineCount = mcp.status?.machineCount ?? 0;
-                const readyCount = mcp.status?.readyMachineCount ?? 0;
-                const updatedCount = mcp.status?.updatedMachineCount ?? 0;
-                const currentConfig = mcp.status?.configuration?.name || '';
-
-                return (
-                  <button key={mcp.metadata.uid} onClick={() => go(`/r/machineconfiguration.openshift.io~v1~machineconfigpools/_/${mcp.metadata.name}`, mcp.metadata.name)}
-                    className="w-full px-4 py-3 hover:bg-slate-800/50 text-left transition-colors">
-                    <div className="flex items-center justify-between mb-1">
-                      <div className="flex items-center gap-2">
-                        <span className="text-sm font-medium text-slate-200">{mcp.metadata.name}</span>
-                        {isDegraded && <span className="text-xs px-1.5 py-0.5 bg-red-900/50 text-red-300 rounded">Degraded</span>}
-                        {isUpdating && <span className="text-xs px-1.5 py-0.5 bg-blue-900/50 text-blue-300 rounded">Updating</span>}
-                        {isUpdated && !isUpdating && <span className="text-xs px-1.5 py-0.5 bg-green-900/50 text-green-300 rounded">Updated</span>}
-                      </div>
-                      <span className={cn('text-xs font-mono', readyCount === machineCount ? 'text-green-400' : 'text-yellow-400')}>
-                        {readyCount}/{machineCount} ready
-                      </span>
-                    </div>
-                    <div className="flex items-center gap-3 text-xs text-slate-500">
-                      <span>{updatedCount}/{machineCount} updated</span>
-                      {currentConfig && <span className="font-mono truncate max-w-[200px]">Config: {currentConfig}</span>}
-                    </div>
-                  </button>
-                );
-              })}
-            </div>
-          </Card>
-
-          {/* MachineConfigs quick access */}
-          <Card>
-            <div className="px-4 py-3 border-b border-slate-800 flex items-center justify-between">
-              <h2 className="text-sm font-semibold text-slate-100">Machine Configuration</h2>
-              <button onClick={() => go('/r/machineconfiguration.openshift.io~v1~machineconfigs', 'MachineConfigs')} className="text-xs text-blue-400 hover:text-blue-300">View all →</button>
-            </div>
-            <div className="p-4 space-y-2">
-              <p className="text-xs text-slate-400">MachineConfigs define OS-level configuration for nodes — systemd units, kernel parameters, files, and more. Changes trigger rolling reboots via MachineConfigPools.</p>
-              <div className="space-y-1.5 pt-2">
-                {[
-                  { label: 'MachineConfigs', path: '/r/machineconfiguration.openshift.io~v1~machineconfigs', desc: 'OS-level node configuration (files, systemd, kernel args)' },
-                  { label: 'MachineConfigPools', path: '/r/machineconfiguration.openshift.io~v1~machineconfigpools', desc: 'Groups of nodes that share the same MachineConfig' },
-                  { label: 'KubeletConfigs', path: '/r/machineconfiguration.openshift.io~v1~kubeletconfigs', desc: 'Kubelet parameters (maxPods, eviction thresholds)' },
-                  { label: 'ContainerRuntimeConfigs', path: '/r/machineconfiguration.openshift.io~v1~containerruntimeconfigs', desc: 'CRI-O runtime settings (pids limit, log size)' },
-                ].map((item) => (
-                  <button key={item.label} onClick={() => go(item.path, item.label)}
-                    className="flex items-center justify-between w-full py-2 px-3 rounded hover:bg-slate-800/50 text-left transition-colors">
-                    <div>
-                      <div className="text-sm text-slate-200">{item.label}</div>
-                      <div className="text-xs text-slate-500">{item.desc}</div>
-                    </div>
-                    <ArrowRight className="w-3 h-3 text-slate-600 shrink-0" />
-                  </button>
-                ))}
-              </div>
-            </div>
-          </Card>
-        </div>
+        <MachineConfigSection machineConfigPools={machineConfigPools} go={go} />
         </>}
       </div>
-    </div>
-  );
-}
-
-function Tip({ title, desc }: { title: string; desc: string }) {
-  return (
-    <div className="flex gap-2">
-      <div className="w-1.5 h-1.5 rounded-full bg-blue-500 mt-1.5 shrink-0" />
-      <div>
-        <span className="font-medium text-slate-200">{title}</span>
-        <p className="text-slate-500">{desc}</p>
-      </div>
-    </div>
-  );
-}
-
-function StatCard({ label, value, subtitle, issues, bar, barColor, onClick }: {
-  label: string; value: string; subtitle?: string; issues?: number; bar?: number | null; barColor?: 'green' | 'yellow' | 'red' | 'blue'; onClick?: () => void;
-}) {
-  return (
-    <div onClick={onClick} className={cn('bg-slate-900 rounded-lg border p-3', onClick && 'cursor-pointer hover:border-slate-600', issues ? 'border-yellow-800' : 'border-slate-800')}>
-      <div className="flex items-center justify-between mb-1">
-        <span className="text-xs text-slate-400">{label}</span>
-        {issues ? <span className="text-xs px-1.5 py-0.5 bg-red-900 text-red-300 rounded">{issues}</span> : null}
-      </div>
-      <div className="text-xl font-bold text-slate-100">{value}</div>
-      {subtitle && <div className="text-xs text-slate-500 mt-0.5">{subtitle}</div>}
-      {bar !== null && bar !== undefined && <UsageBar pct={bar} color={barColor || 'blue'} className="mt-2" />}
-    </div>
-  );
-}
-
-function UsageBar({ pct, color, className }: { pct: number; color: 'green' | 'yellow' | 'red' | 'blue'; className?: string }) {
-  const bg = { green: 'bg-green-500', yellow: 'bg-yellow-500', red: 'bg-red-500', blue: 'bg-blue-500' }[color];
-  return (
-    <div className={cn('w-16 h-1.5 bg-slate-700 rounded-full overflow-hidden', className)}>
-      <div className={cn('h-full rounded-full transition-all', bg)} style={{ width: `${Math.min(100, pct)}%` }} />
     </div>
   );
 }
